Refactor to comply with Microsoft requirements: 1.       The template does not have a baseURL parameter and uses [uri(deployment().properties.templateLink.uri, concat('storageAccount-', parameters('storageAccountNewOrExisting'), '.json'))] to determine the location of linked templates, it would be better if the first part of this was used as the default for baseURL and then the baseURL was used to construct the valriable.
2.       The template can be simplified a lot by using loops, the NICs, VMs, load balancers, IP addresses etc. are all repeated 3 times.

3.       Is it necessary to have RDP on port 3389 for each VM? Two  suggestions here:
Have the first VM only exposed and then use it as a jumpbox to get to the others.
Have one LB and use different ports to expose RDP on the other VMs.

If either of these can be done then the DNS selection can be done using the PublicIP combo control (rather than concatenating unique string on to the end of a user supplied string) and there will be validation in the portal

4.       Storage wise in createUIDefinition and the two storage account templates it needs to exclude the use of Premium Storage,  the allowed VM sizes do not support it